refactor(bridge): extract structural action diagnostics
